Noah's edits to Saturday changes
I readded the deployCode file. As far as I'm aware, it's the easiest way to upload changes from our computers to the raspberry pi. I know git is now installed on the pi, but we may be on a WIFI network that blocks GitHub. Let's keep it just in case.

I also renamed rgbmatrix.py back to what is used to be. It needs to be this name for simulation to work. If it causes issues on the raspberry pi, just delete it off of there.

I also reworked most of the networktables stuff. I think we need a meeting to discuss how networktables works.